+++ src/App.tsx
@@ -356,8 +356,6 @@ export class App extends React.Component<AppProps, AppState> {
                 searchResults: [found.match],
               })
               break
-            case "batch":
-              throw new Error("unreachable")
             default:
               assertNever(found)
           }
@@ -596,8 +594,6 @@ export class App extends React.Component<AppProps, AppState> {
               case "found":
                 changes.searchResults = [found.match]
                 break
-              case "batch":
-                throw new Error("unreachable")
               default:
                 assertNever(found)
             }