+++ src/Search.tsx
@@ -247,8 +247,6 @@ function ResultsInfo({
               case "found":
                 searchResults = [results.match]
                 break
-              case "batch":
-                throw new Error("unreachable")
               default:
                 assertNever(results)
             }
@@ -276,8 +274,6 @@ function ResultsInfo({
               case "found":
                 searchResults = [results.match]
                 break
-              case "batch":
-                throw new Error("unreachable")
               default:
                 assertNever(results)
             }
@@ -441,8 +437,6 @@ function Form({ app, resetter }: { app: App; resetter: () => void }) {
       search = { type: "ad hoc", phrase: app.state.search.phrase }
       app.setState({ search })
       break
-    case "batch":
-      throw new Error("unreachable")
     default:
       search = deepClone(app.state.search)
       break
@@ -555,8 +549,6 @@ function Form({ app, resetter }: { app: App; resetter: () => void }) {
                   case "ambiguous":
                     searchResults = results.matches
                     break
-                  case "batch":
-                    throw new Error("unreachable")
                   default:
                     assertNever(results)
                 }
@@ -806,8 +798,6 @@ function Form({ app, resetter }: { app: App; resetter: () => void }) {
                           reset(search)
                         )
                         break
-                      case "batch":
-                        throw new Error("unreachable")
                       default:
                         assertNever(found)
                     }