+++ CHANGES.md
@@ -5,6 +5,7 @@ Key changes in each release. The dates with each release are the dates I submitt
 *0.0.5 pending*
 - minor interface changes
 - improve known word in gist lookup so it now can handle phrases
+- various efficiency improvements
 
 *0.0.4 2022-1-2*
 - added forward and back hotkeys for moving about in note history